Create devcontainer.json
Features
Dev Container Features are self-contained units of installation code and development container configuration. Features are designed to install atop a wide-range of base container images. Be sure to check a Feature's documentation to see which base images are supported.

Usage
To reference a Feature, add the desired Feature(s) to a devcontainer.json. See a Feature's README for available options, or leave the options empty to use the default value.

Example devcontainer.json
{
    "name": "my-project-devcontainer",
    "image": "mcr.microsoft.com/devcontainers/base:ubuntu",  // Any generic, debian-based image.
    "features": {
        "ghcr.io/devcontainers/features/go:1": {
            "version": "1.18"
        },
        "ghcr.io/devcontainers/features/docker-in-docker:1": {
            "version": "latest",
            "moby": true
        }
    }
}